move actionTemplates out of config
