+++ src/config/index.ts
@@ -42,4 +42,3 @@ export {
   SUPPORTED_TRANSACTION_QUERY_PARAMS,
   MANDATORY_TRANSACTION_QUERY_PARAMS
 } from './queryParams';
-export { actionTemplates } from './userActions';
